Make command file for chaoslib
