--- pkg/kube-aws/ebs-loss-by-id/environment/environment.go
@@ -1,58 +1,0 @@
-package environment
-
-import (
-	"os"
-	"strconv"
-
-	clientTypes "k8s.io/apimachinery/pkg/types"
-
-	experimentTypes "github.com/litmuschaos/litmus-go/pkg/kube-aws/ebs-loss-by-id/types"
-	"github.com/litmuschaos/litmus-go/pkg/types"
-)
-
-//GetENV fetches all the env variables from the runner pod
-func GetENV(experimentDetails *experimentTypes.ExperimentDetails) {
-	experimentDetails.ExperimentName = Getenv("EXPERIMENT_NAME", "ebs-loss")
-	experimentDetails.ChaosNamespace = Getenv("CHAOS_NAMESPACE", "litmus")
-	experimentDetails.EngineName = Getenv("CHAOSENGINE", "")
-	experimentDetails.ChaosDuration, _ = strconv.Atoi(Getenv("TOTAL_CHAOS_DURATION", "30"))
-	experimentDetails.ChaosInterval, _ = strconv.Atoi(Getenv("CHAOS_INTERVAL", "30"))
-	experimentDetails.RampTime, _ = strconv.Atoi(Getenv("RAMP_TIME", "0"))
-	experimentDetails.ChaosLib = Getenv("LIB", "litmus")
-	experimentDetails.AppNS = Getenv("APP_NAMESPACE", "")
-	experimentDetails.AppLabel = Getenv("APP_LABEL", "")
-	experimentDetails.AppKind = Getenv("APP_KIND", "")
-	experimentDetails.ChaosUID = clientTypes.UID(Getenv("CHAOS_UID", ""))
-	experimentDetails.InstanceID = Getenv("INSTANCE_ID", "")
-	experimentDetails.ChaosPodName = Getenv("POD_NAME", "")
-	experimentDetails.AuxiliaryAppInfo = Getenv("AUXILIARY_APPINFO", "")
-	experimentDetails.Delay, _ = strconv.Atoi(Getenv("STATUS_CHECK_DELAY", "2"))
-	experimentDetails.Timeout, _ = strconv.Atoi(Getenv("STATUS_CHECK_TIMEOUT", "180"))
-	experimentDetails.EBSVolumeID = Getenv("EBS_VOLUME_ID", "")
-	experimentDetails.Region = Getenv("REGION", "")
-	experimentDetails.Sequence = Getenv("SEQUENCE", "parallel")
-	experimentDetails.TargetContainer = Getenv("TARGET_CONTAINER", "")
-}
-
-// Getenv fetch the env and set the default value, if any
-func Getenv(key string, defaultValue string) string {
-	value := os.Getenv(key)
-	if value == "" {
-		value = defaultValue
-	}
-	return value
-}
-
-//InitialiseChaosVariables initialise all the global variables
-func InitialiseChaosVariables(chaosDetails *types.ChaosDetails, experimentDetails *experimentTypes.ExperimentDetails) {
-
-	chaosDetails.ChaosNamespace = experimentDetails.ChaosNamespace
-	chaosDetails.ChaosPodName = experimentDetails.ChaosPodName
-	chaosDetails.ChaosUID = experimentDetails.ChaosUID
-	chaosDetails.EngineName = experimentDetails.EngineName
-	chaosDetails.ExperimentName = experimentDetails.ExperimentName
-	chaosDetails.InstanceID = experimentDetails.InstanceID
-	chaosDetails.Timeout = experimentDetails.Timeout
-	chaosDetails.Delay = experimentDetails.Delay
-	chaosDetails.ProbeImagePullPolicy = experimentDetails.LIBImagePullPolicy
-}
